searchapi/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
fastapi==0.112.0
filelock==3.15.4
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.5
idna==3.7
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
logging==0.4.9.6
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
openai==1.40.2
packaging==24.1
pillow==10.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pytesseract==0.3.13
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
safetensors==0.4.4
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.1.1
sniffio==1.3.1
soupsieve==2.6
starlette==0.37.2
sympy==1.13.2
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tqdm==4.66.5
transformers==4.44.0
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.5
